# minute-main/openapi-temp.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/check-access": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Check Admin Access",
        "description": "Check if current user has admin access.",
        "operationId": "check_admin_access_admin_check_access_get",
        "parameters": [
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/configs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Configs",
        "description": "List all available tenant configurations.",
        "operationId": "list_configs_admin_configs_get",
        "parameters": [
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/configs/{tenant_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Config Detail",
        "description": "Get detailed configuration for a specific tenant.",
        "operationId": "get_config_detail_admin_configs__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/configs/{tenant_id}/audit": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Config Audit",
        "description": "Get audit trail for a specific tenant configuration.",
        "operationId": "get_config_audit_admin_configs__tenant_id__audit_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigAuditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/configs/{tenant_id}/history": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Config History",
        "description": "Get version history for a specific tenant configuration.",
        "operationId": "get_config_history_admin_configs__tenant_id__history_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessCheckResponse": {
        "properties": {
          "hasAccess": {
            "type": "boolean",
            "title": "Hasaccess"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "hasAccess"
        ],
        "title": "AccessCheckResponse"
      },
...
      "ConfigAuditResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryResponse"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "ConfigAuditResponse"
      },
      "ConfigHistoryResponse": {
        "properties": {
          "configName": {
            "type": "string",
            "title": "Configname"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ConfigVersionItem"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "configName",
          "versions"
        ],
        "title": "ConfigHistoryResponse"
      },
      "ConfigListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "lastModified": {
            "type": "string",
            "title": "Lastmodified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "lastModified"
        ],
        "title": "ConfigListItem"
      },
      "ConfigListResponse": {
        "properties": {
          "configs": {
            "items": {
              "$ref": "#/components/schemas/ConfigListItem"
            },
            "type": "array",
            "title": "Configs"
          }
        },
        "type": "object",
        "required": [
          "configs"
        ],
        "title": "ConfigListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LanguageConfig": {
        "properties": {
          "default": {
            "type": "string",
            "title": "Default",
            "description": "Default language code (e.g. en)"
          },
          "available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available",
            "description": "List of languages that can be requested"
          },
          "autoTranslate": {
            "type": "boolean",
            "title": "Autotranslate",
            "description": "When true, automatically request translations after transcription completes",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "default"
        ],
        "title": "LanguageConfig"
      },
...
      "TenantConfig": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable tenant/council identifier"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Config version tag, e.g. 1.0.0"
          },
          "defaultLocale": {
            "type": "string",
            "pattern": "^[a-z]{2}-[A-Z]{2}$",
            "title": "Defaultlocale"
          },
          "designTokens": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Designtokens"
          },
          "retentionDaysDefault": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retentiondaysdefault",
            "description": "Default retention in days for recordings/transcriptions/minutes"
          },
          "sharepoint_library_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharepoint Library Path",
            "description": "Default SharePoint library path"
          },
          "planner": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planner",
            "description": "Optional Planner plan/bucket ids"
          },
          "languages": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Language/translation configuration for the tenant"
          },
          "navigation": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NavigationItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Navigation",
            "description": "Optional navigation overrides"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleConfig"
            },
            "type": "array",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "defaultLocale",
          "modules"
        ],
        "title": "TenantConfig"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}